update samples_played for AAC seeking and fix 32 bit integer overflow in positionMillis

+++ codecs.h
@@ -103,7 +103,7 @@ public:
 	virtual void stop() = 0;
 	virtual int play() = 0;
 	bool isPlaying(void) {return playing > 0;}
-	unsigned positionMillis(void) { return samples_played * 1000 / AUDIOCODECS_SAMPLE_RATE;}
+	unsigned positionMillis(void) { return ((uint64_t)samples_played) * 1000 / AUDIOCODECS_SAMPLE_RATE;}
 	unsigned lengthMillis(void) {return max(fsize() / (bitrate / 8 ) * 1000,  positionMillis());} //Ignores VBR
 	int channels(void) {return _channels;}
 	int bitRate(void) {return bitrate;}
